[Refactor] #14: 패키지 앞글자 소문자로 변경
