+++ src/main/java/com/example/BarrierKU/domain/building/repository/BuildingRepository.java
@@ -1,6 +1,6 @@
 package com.example.BarrierKU.domain.building.repository;
 
-import com.example.BarrierKU.domain.Indoor.Building;
+import com.example.BarrierKU.domain.indoor.Building;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BuildingRepository extends JpaRepository<Building, Long> {
