+++ src/main/java/com/example/BarrierKU/domain/type/ImageType.java
@@ -1,4 +1,4 @@
-package com.example.BarrierKU.domain.Type;
+package com.example.BarrierKU.domain.type;
 
 public enum ImageType {
     DOOR,
